- Fix dto classes

+++ src/server/capstone/src/main/java/com/sb/brothers/capstone/dto/PostDto.java
@@ -31,7 +31,7 @@ public class PostDto {
     //private String modifiedBy;
     private int status;
     private int noDays;
-    private double fee;
+    private int fee;
     //private String manager;
     private String user;
     private List<PostDetailDto> postDetailDtos;
